PdfAfent/requirements.txt
altair==4.2.2
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
distro==1.9.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.17.0
numpy==2.0.2
openai==2.24.0
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
protobuf==3.20.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rich==14.3.3
rpds-py==0.27.1
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.12.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
validators==0.35.0
zipp==3.23.0
